refactor: 수량 상태 업데이트 로직 커스텀훅으로 분리
